[Ricardo] updated readme with repeatable steps, added application properties. parameterized deploy script for emr public dns

+++ deploy.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
-sbt assembly
-FILE=target.zip
-if test -f "$FILE"; then
-    rm -rf $FILE
+
+if $# -gt 0; then
+  location=$1
+
+  sbt assembly
+  FILE=target.zip
+  if test -f "$FILE"; then
+      rm -rf $FILE
+  fi
+  zip -r target.zip target
+  scp -i EMR-key-pair.pem target.zip hadoop@"$location":~
+
+  else
+    echo "Error: no EMR public master dns found"
+    echo "example: ./deploy.sh ec2.example.com"
 fi
-zip -r target.zip target
-scp -i EMR-key-pair.pem target.zip hadoop@ec2-3-22-70-62.us-east-2.compute.amazonaws.com:~
